make author cell type params optional and autodetect delimiters
