+++ .gitignore
@@ -93,3 +93,4 @@ target/
 profile_default/
 ipython_config.py
 
+test_human/
