+++ bin/get_census_adata.py
@@ -45,11 +45,11 @@ def parse_arguments():
     parser.add_argument('--organ', type=str, default="brain")
     parser.add_argument('--assay', type=str, nargs = "+", help="Assays to subset from referenc (unnecessary)", default=None)
     parser.add_argument('--tissue', type=str, nargs="+", default = None, help = "tissues to pull from (different from organ, this can select for more specific brain regions)")
-    parser.add_argument('--subsample', type=int, help="Number of cells per cell type to subsample from reference", default=500)
+    parser.add_argument('--subsample', type=int, help="Number of cells per cell type to subsample from reference", default=50)
     parser.add_argument('--rename_file', type=str, default="/space/grp/rschwartz/rschwartz/cell_annotation_cortex.nf/meta/author_cell_annotations/rename_cells_mmus_author.tsv")
     parser.add_argument('--ref_name', type=str, default="whole_cortex", help="Prefix of temporary reference file created")
-    parser.add_argument('--original_celltype_columns', type=str, default="/space/grp/rschwartz/rschwartz/cell_annotation_cortex.nf/meta/author_cell_annotations/original_celltype_columns.tsv")
-    parser.add_argument('--author_annotations_path', type=str, default="/space/grp/rschwartz/rschwartz/cell_annotation_cortex.nf/meta/author_cell_annotations")
+    parser.add_argument('--original_celltype_columns', type=str, default=None)
+    parser.add_argument('--author_annotations_path', type=str, default=None)
     
     if __name__ == "__main__":
         known_args, _ = parser.parse_known_args()
@@ -73,7 +73,8 @@ def main():
    original_celltype_columns = args.original_celltype_columns
    author_annotations_path = args.author_annotations_path
    
-   if organism == "mus_musculus":
+   #f organism == "mus_musculus":
+   if original_celltype_columns is not None:
       original_celltypes = get_original_celltypes(columns_file=original_celltype_columns,
                                           author_annotations_path=author_annotations_path) 
    else:
