+++ nextflow.config
@@ -4,25 +4,22 @@ manifest {
     mainScript = 'main.nf'
     version = '1.1.0'
 }
-
-params.run_download = false
-params.studies_path = "/home/rxiang/rachelpipeline/human/"
+params.study_names = null
+params.studies_path = "/home/rxiang/rachelpipeline/mouse/"
 params.gene_mapping = "$projectDir/meta/gemma_genes.tsv"
-params.original_celltype_columns = "$projectDir/meta/author_cell_annotations/original_celltype_columns.tsv"
-params.author_annotations_path = "$projectDir/meta/author_cell_annotations"
+params.original_celltype_columns = null 
+params.author_annotations_path = null
 params.nmads=5
 params.version = "$manifest.version"
 process.executor = 'local'
 params.GEMMA_USERNAME = System.getenv('GEMMA_USERNAME')
 params.GEMMA_PASSWORD = System.getenv('GEMMA_PASSWORD')
-params.study_names = "study_names.txt"
 // Define the required input parameters
 params.organism = "mus_musculus"
-params.restricted_celltypes = ["unknown", "neuron"]
 params.census_version = "2024-07-01" // version of cellxgene census scvi model and data corpus for reference data
 params.cutoff = 0 // do not threshold class probabilities 
 params.seed = 42
-params.subsample_ref = 50
+params.subsample_ref = 500
 params.ref_collections = [
         "A taxonomy of transcriptomic cell types across the isocortex and hippocampal formation",
         "An integrated transcriptomic and epigenomic atlas of mouse primary motor cortex cell types",
