+++ params.hs.json
@@ -4,11 +4,12 @@
     "cutoff": 0,
     "seed": 42,
     "subsample_ref": 500,
-    "study_names": "study_names_human.txt",
+    "studies_path": "/space/grp/rschwartz/rschwartz/cell_annotation_cortex.nf/test_human",
     "ref_collections": [
         "Transcriptomic cytoarchitecture reveals principles of human neocortex organization",
         "SEA-AD: Seattle Alzheimer’s Disease Brain Cell Atlas"
     ],
     "organ": "brain",
-    "rename_file": "/space/grp/Pipelines/sc-annotation-pipeline/meta/rename_cells_hsap.tsv"
+    "rename_file": "/space/grp/Pipelines/sc-annotation-pipeline/meta/rename_cells_hsap.tsv",
+    "nmads": 5
 }
